# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel CRUD",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "operationId": "7fdb9b40bd4ad2baf12aa15164eab788",
                "responses": {
                    "200": {
                        "description": "Получить содержимое корзины",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 4242
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Swagger Book"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "example": 1800
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cart/current": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "operationId": "681d5171020de9916346545a719c2325",
                "responses": {
                    "200": {
                        "description": "Получить статистику лежащего в корзине",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "articles": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 4242
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Swagger Book"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "example": 1800
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total_sum": {
                                            "type": "number",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cart/{itemId}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "operationId": "61467a79ad43faca035fc4491264d48a",
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "ID товара",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Получить статистику лежащего в корзине",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/community/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "operationId": "7ed22acf70b24517e8851d1fe7f270b2",
                "responses": {
                    "200": {
                        "description": "List of messages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "Lorem ipsum dolot sit amet"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2024-11-30 08:22:13"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Lokusok"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "operationId": "66dc3027b3e0fe7cbbace8e5b20d7794",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Text of the message"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Save message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "Lorem ipsum dolot sit amet"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "example": "2024-11-30 08:22:13"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Lokusok"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/community/messages/{message}": {
            "delete": {
                "tags": [
                    "Messages"
                ],
                "operationId": "e8d43a5c8a8c4f437e44110a845f4b1c",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "description": "ID сообщения для удаления",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Destroy message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success delete"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Cart",
            "description": "Cart"
        },
        {
            "name": "Messages",
            "description": "Messages"
        }
    ]
}